# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mnemos API",
  "description": "Governed long-term memory and continuous learning for AI agents.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/v1/experiences": {
   "post": {
    "tags": [
     "experiences"
    ],
    "summary": "Create Experience",
    "operationId": "create_experience_v1_experiences_post",
    "parameters": [
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExperienceIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExperienceCreated"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "experiences"
    ],
    "summary": "List Experiences",
    "operationId": "list_experiences_v1_experiences_get",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Workspace Id"
      }
     },
     {
      "name": "project_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Project Id"
      }
     },
     {
      "name": "agent_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Agent Id"
      }
     },
     {
      "name": "outcome",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Outcome"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Experiences V1 Experiences Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/experiences/{experience_id}": {
   "get": {
    "tags": [
     "experiences"
    ],
    "summary": "Get Experience",
    "operationId": "get_experience_v1_experiences__experience_id__get",
    "parameters": [
     {
      "name": "experience_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Experience Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExperienceDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/episodes": {
   "get": {
    "tags": [
     "experiences"
    ],
    "summary": "List Episodes",
    "operationId": "list_episodes_v1_episodes_get",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Workspace Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Episodes V1 Episodes Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/episodes/{episode_id}": {
   "get": {
    "tags": [
     "experiences"
    ],
    "summary": "Get Episode",
    "operationId": "get_episode_v1_episodes__episode_id__get",
    "parameters": [
     {
      "name": "episode_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Episode Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EpisodeDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EpisodeDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "task_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Task Id"
     },
     "summary": {
      "type": "string",
      "title": "Summary"
     },
     "outcome": {
      "type": "string",
      "title": "Outcome"
     },
     "importance": {
      "type": "number",
      "title": "Importance"
     },
     "confidence": {
      "type": "number",
      "title": "Confidence"
     },
     "experience_count": {
      "type": "integer",
      "title": "Experience Count"
     },
     "started_at": {
      "type": "string",
      "format": "date-time",
      "title": "Started At"
     },
     "completed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Completed At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "experiences": {
      "items": {
       "$ref": "#/components/schemas/ExperienceOut"
      },
      "type": "array",
      "title": "Experiences"
     }
    },
    "type": "object",
    "required": [
     "id",
     "workspace_id",
     "project_id",
     "agent_id",
     "session_id",
     "task_id",
     "summary",
     "outcome",
     "importance",
     "confidence",
     "experience_count",
     "started_at",
     "completed_at",
     "updated_at",
     "experiences"
    ],
    "title": "EpisodeDetail"
   },
...
   "ExperienceCreated": {
    "properties": {
     "experience": {
      "$ref": "#/components/schemas/ExperienceOut"
     },
     "learning": {
      "$ref": "#/components/schemas/LearningStatus"
     }
    },
    "type": "object",
    "required": [
     "experience",
     "learning"
    ],
    "title": "ExperienceCreated"
   },
   "ExperienceDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "task_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Task Id"
     },
     "episode_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Episode Id"
     },
     "task": {
      "type": "string",
      "title": "Task"
     },
     "observation": {
      "type": "string",
      "title": "Observation"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "result": {
      "type": "string",
      "title": "Result"
     },
     "outcome": {
      "type": "string",
      "title": "Outcome"
     },
     "importance": {
      "type": "number",
      "title": "Importance"
     },
     "confidence": {
      "type": "number",
      "title": "Confidence"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "source_trust": {
      "type": "number",
      "title": "Source Trust"
     },
     "metadata_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metadata Json"
     },
     "processing_status": {
      "type": "string",
      "title": "Processing Status"
     },
     "processed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Processed At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "learning": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/LearningStatus"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "workspace_id",
     "project_id",
     "agent_id",
     "session_id",
     "task_id",
     "episode_id",
     "task",
     "observation",
     "action",
     "result",
     "outcome",
     "importance",
     "confidence",
     "source",
     "source_trust",
     "metadata_json",
     "processing_status",
     "processed_at",
     "created_at"
    ],
    "title": "ExperienceDetail"
   },
   "ExperienceIn": {
    "properties": {
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "project_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Name"
     },
     "agent_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Name"
     },
     "task_id": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Task Id"
     },
     "task": {
      "type": "string",
      "maxLength": 20000,
      "minLength": 1,
      "title": "Task"
     },
     "observation": {
      "type": "string",
      "maxLength": 20000,
      "title": "Observation",
      "default": ""
     },
     "action": {
      "type": "string",
      "maxLength": 20000,
      "title": "Action",
      "default": ""
     },
     "result": {
      "type": "string",
      "maxLength": 20000,
      "title": "Result",
      "default": ""
     },
     "outcome": {
      "$ref": "#/components/schemas/Outcome"
     },
     "importance": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Importance",
      "default": 0.5
     },
     "confidence": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Confidence",
      "default": 0.7
     },
     "source": {
      "$ref": "#/components/schemas/ExperienceSource",
      "default": "agent"
     },
     "metadata": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metadata"
     }
    },
    "type": "object",
    "required": [
     "workspace_id",
     "task",
     "outcome"
    ],
    "title": "ExperienceIn"
   },
   "ExperienceOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "task_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Task Id"
     },
     "episode_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Episode Id"
     },
     "task": {
      "type": "string",
      "title": "Task"
     },
     "observation": {
      "type": "string",
      "title": "Observation"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "result": {
      "type": "string",
      "title": "Result"
     },
     "outcome": {
      "type": "string",
      "title": "Outcome"
     },
     "importance": {
      "type": "number",
      "title": "Importance"
     },
     "confidence": {
      "type": "number",
      "title": "Confidence"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "source_trust": {
      "type": "number",
      "title": "Source Trust"
     },
     "metadata_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metadata Json"
     },
     "processing_status": {
      "type": "string",
      "title": "Processing Status"
     },
     "processed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Processed At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "workspace_id",
     "project_id",
     "agent_id",
     "session_id",
     "task_id",
     "episode_id",
     "task",
     "observation",
     "action",
     "result",
     "outcome",
     "importance",
     "confidence",
     "source",
     "source_trust",
     "metadata_json",
     "processing_status",
     "processed_at",
     "created_at"
    ],
    "title": "ExperienceOut"
   },
   "ExperienceSource": {
    "type": "string",
    "enum": [
     "agent",
     "user",
     "tool",
     "external"
    ],
    "title": "ExperienceSource"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "LearningStatus": {
    "properties": {
     "job_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Id"
     },
     "job_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Status"
     },
     "processing_status": {
      "type": "string",
      "title": "Processing Status"
     },
     "memories": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Memories"
     }
    },
    "type": "object",
    "required": [
     "job_id",
     "job_status",
     "processing_status",
     "memories"
    ],
    "title": "LearningStatus"
   },
...
   "Outcome": {
    "type": "string",
    "enum": [
     "success",
     "failure",
     "partial",
     "unknown"
    ],
    "title": "Outcome"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}